Move custom extractor test to test of _adapt_url
